app/requirements.txt
fastapi[all]
uvicorn[standard]
pytesseract
Pillow
pdf2image
python-multipart
